move all metrics from pool_synctime_metric file to metrics.go and remove pool_synctime_metrics.go
